calimero_storage_macros
Derive Macro AtomicUnit

#[derive(AtomicUnit)]
{
    // Attributes available to this derive:
    #[children]
    #[collection]
    #[private]
    #[root]
    #[skip]
    #[storage]
    #[type_id]
}

Derives the AtomicUnit trait for a struct.

This macro automatically implements the AtomicUnit trait for a struct, which extends the Data trait.

Requirements

The following are mandatory requirements for the struct:

A private storage field of type Element. This is needed as the Data-based struct needs to own an Element.

Generated implementations

This macro will generate the following implementations:

Data trait implementation.

AtomicUnit trait implementation.

Getter and setter methods for each field. These help to ensure that the access to the fields is controlled, and that any changes to the fields are reflected in the Element’s state.

BorshSerialize and BorshDeserialize will be implemented for the struct, so they should be omitted from the struct definition.

Struct attributes

#[root] - Indicates that the type represents a root in the hierarchy, and doesn’t have a parent. This is an optional attribute.

#[type_id(n)] - Indicates the type ID for the struct. This is a mandatory attribute, and the value n must be a u8. This is used to differentiate between different types of structs when deserialising, and each type should have a unique ID.

Field attributes

#[collection] - Indicates that the field is a collection of other Data types.

#[private] - Designates fields that are local-only, and so should not be shared with other nodes in the network. These fields will not be serialised or included in the Merkle hash calculation. Note that being local-only is not the same as applying permissions via ACLs to share with only the current user — these fields are not shared at all.

#[skip] - Can be applied to fields that should not be serialised or included in the Merkle hash calculation. These fields will be completely ignored by the storage system, and not even have getters and setters implemented.

#[storage] - Indicates that the field is the storage element for the struct. This is a mandatory field, and if it is missing, there will be a panic during compilation. The name is arbitrary, but the type has to be an Element.

Note that fields marked with #[private] or #[skip] must have Default implemented so that they can be initialised when deserialising.

TODO: The #[private] attribute is implemented with the same functionality as #[skip], but in future these will be differentiated.

Getters and setters

The macro will generate getter and setter methods for each field. These methods will allow the struct to control access to its fields, and ensure that any changes to the fields are reflected in the Element’s state.

The getter methods will have the same name as the field, and the setter methods will be prefixed with set_. For example, given a field name, the getter method will be name(), and the setter method will be set_name().

The setter methods will return a boolean indicating whether the update was carried out. Note that this is more an indication of change than it is of error — if the value is the same as the current value, the update will not be carried out, and the method will return false.

Examples

use calimero_storage::entities::Element;
use calimero_storage_macros::AtomicUnit;
use borsh::{BorshSerialize, BorshDeserialize};

#[derive(AtomicUnit, Clone, Debug, Eq, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
#[type_id(43)]
struct Page {
    title: String,
    #[private]
    secret: String,
    #[storage]
    storage: Element,
}

use calimero_storage::entities::Element;
use calimero_storage_macros::{AtomicUnit, Collection};
use borsh::{BorshSerialize, BorshDeserialize};

#[derive(AtomicUnit, Clone, Debug, Eq, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
#[type_id(44)]
struct Person {
    name: String,
    age: u32,
    #[private]
    secret: String,
    #[collection]
    friends: Friends,
    #[storage]
    storage: Element,
}

#[derive(Collection, Clone, Debug, Eq, PartialEq, PartialOrd)]
#[children(Person)]
struct Friends;

Panics

This macro will panic during compilation if:

It is applied to anything other than a struct

The struct has unnamed fields

The struct does not have a field annotated as #[storage]

The struct has fields with types that do not implement Default

The struct already has methods with the same names as the generated getter and setter methods

See also

Collection - For defining a collection of child elements, for use with AtomicUnit (the parent and children are all atomic units, with the collection being the grouping mechanism at field level on the parent).