calimero_storage_macros
Derive Macro Collection

#[derive(Collection)]
{
    // Attributes available to this derive:
    #[children]
}

Derives the Collection trait for a struct.

This macro will automatically implement the Collection trait for the struct it’s applied to.

Requirements

The following are mandatory requirements for the struct:

A #[children(Type)] attribute to specify the type of the children in the Collection.

Generated implementations

This macro will generate the following implementations:

Collection trait implementation.

BorshSerialize and BorshDeserialize will be implemented for the struct, so they should be omitted from the struct definition.

Struct attributes

#[children] - A mandatory attribute to specify the child type for the struct, written as #[children(ChildType)]. Neither the attribute nor its value can be omitted.

Field attributes

None.

Examples

use calimero_storage_macros::{AtomicUnit, Collection};
use calimero_storage::entities::{Data, Element};
use borsh::{BorshSerialize, BorshDeserialize};

#[derive(AtomicUnit, Clone, Debug, Eq, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
struct Book {
    title: String,
    pages: Pages,
    #[storage]
    storage: Element,
}

#[derive(Collection, Clone, Debug, Eq, PartialEq, PartialOrd)]
#[children(Page)]
struct Pages;

#[derive(AtomicUnit, Clone, Debug, Eq, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
struct Page {
    content: String,
    #[storage]
    storage: Element,
}

Panics

This macro will panic during compilation if:

It is applied to anything other than a struct

The struct has unnamed fields

The #[children(Type)] attribute is missing or invalid

See also

AtomicUnit - For defining a single atomic unit of data that either stands alone, or owns one or more collections, or is a child in a collection.